refactor: split monolithic resolvers.test.ts into domain-specific files
Split the 4,541-line resolvers.test.ts into 9 focused test files:
- resolvers-admin.test.ts (15 tests) - Admin queries, dashboard, email, errors
- resolvers-auth.test.ts (16 tests) - Authentication mutations
- resolvers-crud.test.ts (21 tests) - Comment, Fact, LifeEvent, Source CRUD
- resolvers-family.test.ts (44 tests) - Family queries, mutations, type resolvers
- resolvers-media.test.ts (11 tests) - Media mutations and type resolvers
- resolvers-misc.test.ts (15 tests) - Stats, crests, surname crest mutations
- resolvers-person.test.ts (56 tests) - Person queries, mutations, type resolvers
- resolvers-research.test.ts (18 tests) - Research queue, status, tips, caching
- resolvers-user.test.ts (22 tests) - User mutations and type resolvers

All 393 tests pass. Consistent file structure template across all files.
